InnoDB compiler warning fixes for all Windows builds. This patch uses C++ style casts like static_cast<>(). Approved in rb4338 by Vasil

+++ storage/innobase/fts/fts0sql.cc
@@ -117,7 +117,7 @@ fts_get_table_name_prefix(
 
 	if (slash) {
 		/* Print up to and including the separator. */
-		dbname_len = (slash - fts_table->parent) + 1;
+		dbname_len = static_cast<int>(slash - fts_table->parent) + 1;
 	}
 
 	len = fts_get_table_id(fts_table, table_id);
@@ -152,7 +152,8 @@ fts_get_table_name(
 
 	prefix_name = fts_get_table_name_prefix(fts_table);
 
-	name_len = strlen(prefix_name) + 1 + strlen(fts_table->suffix) + 1;
+	name_len = static_cast<int>(
+		strlen(prefix_name) + 1 + strlen(fts_table->suffix) + 1);
 
 	name = static_cast<char*>(mem_alloc(name_len));
